+++ src/arch/mod.rs
@@ -1,4 +1,5 @@
 pub mod catch;
+pub mod cpu;
 pub mod gdt;
 pub mod gs;
 pub mod idt;